[idle] Fix the bug transitioning from active -> locked -> idle on android.
On Android, the screen gets locked as soon as it gets powered down. So,
unlike Desktop OSes where the screen saver kicks first (which allows us
to detect idleness) and the screen lock second (throwing a "locked"
state), on Android, the "idle" state happens before the "locked" state.

Isolates the bug we are experiencing on Android on a unittest so we can
assert that this behavior willl be part of the API and wont regress.

Bug: 878979
Change-Id: I7b4617dde86609311e10e3bae5a0b52f9db991a7