domparsing: Fix an issue that XMLSerializer generates new prefix for an attribute with an existing namespace
Add an implementation of "retrieving a preferred prefix string"
[1], and apply it to attribute serialization.

The specification says we need to keep a map like [2]

  namespace URI => [older prefix, ..., newer prefix]

in order to retrieve most-recently-declared prefix. However, our
current implementation is:

  prefix => (namespace URI, declaration index)

When we search the map for a namespace URI, we chose a prefix with
the largest declaration index.

The failing test case in XMLSerializer-serializeToString.html follows
the specification, but Firefox, Safari, and Chrome without this CL
preserve prefix in such case.  This CL keeps the interoperable behavior.

[1] https://w3c.github.io/DOM-Parsing/#dfn-retrieving-a-preferred-prefix-string
[2] https://w3c.github.io/DOM-Parsing/#dfn-namespace-prefix-map

Bug: 906807
Change-Id: Idaea590ba19c9c21acbaf456ea91ce550c74db5a